Improve fetching cache from chained backend
refs #19

While profiling a production server noticed during a tracking request which takes like 50ms we spent 2.73 ms in `doContains` and 2ms in `doFetch`. Considering this is like 4% of the time spent of a tracking request in `doContains()` even though it's not needed. The other backends will silently fail anyway and return false when a cache is not contained in a backend. And sometimes they do the `doContains` in `doFetch` anyway (eg ArrayCache).

FYI @mattab