Catch a parse error in case cache file does not get written correctly
Avoids getting a message like this in the UI:

`syntax error, unexpected '=>' (T_DOUBLE_ARROW), expecting ')'`

Instead we return false so the cache gets written again. Not sure if otherwise it would even leak potential data as it shows part of the broken part of the cache. Also it makes the app more stable.